fix minor things; cleanup
