+++ src/continuity/data/function/__init__.py
@@ -1,5 +1,6 @@
 from .function import Function, ParameterizedFunction
-from .function_set import FunctionSet, SampledFunctionSet
+from .function_set import FunctionSet
+from .sampled_function_set import SampledFunctionSet
 
 __all__ = [
     "Function",
